feat 钉钉消息通知

+++ bootx-common-starters/common-starter-dingtalk/src/main/java/cn/bootx/starter/dingtalk/configuration/DingTalkProperties.java
@@ -21,6 +21,6 @@ public class DingTalkProperties {
     private String appSecret;
 
     /** 应用id */
-    private Integer agentId;
+    private Long agentId;
 
 }
